+++ src/services/StateService.ts
@@ -1,11 +1,13 @@
 import { RootState } from "redux/store";
 import { AuthService } from "./AuthService";
+import { IBoredAPIService, BoredAPIService } from "./BoredAPIService";
 import { UserService } from "./UserService";
 
 interface IStateServices {
+	_store: RootState;
 	auth: AuthService;
 	user: UserService;
-	_store: RootState;
+	boredAPI: IBoredAPIService;
 }
 
 /**
@@ -15,11 +17,13 @@ class StateService implements IStateServices {
 	_store = null;
 	auth = null;
 	user = null;
+	boredAPI = null;
 
 	constructor(reduxStore) {
 		this._store = reduxStore;
 		this.auth = new AuthService(reduxStore);
 		this.user = new UserService(reduxStore);
+		this.boredAPI = new BoredAPIService();
 	}
 }